Move session sync logic to PageWithUser
